fix overflow issue

+++ 0063-Unique-Paths-II/cpp-0063/main.cpp
@@ -22,7 +22,7 @@ public:
         if(!n || obstacleGrid[0][0])
             return 0;
 
-        vector<vector<long long>> dp(m, vector<int>(n, -1));
+        vector<vector<long long>> dp(m, vector<long>(n, -1));
         dp[0][0] = 1;
         for(int j = 1; j < n; j ++)
             if(obstacleGrid[0][j])
@@ -42,7 +42,7 @@ public:
                     dp[i][j] = 0;
                 else
                     dp[i][j] = dp[i - 1][j] + dp[i][j - 1];
-        return dp[m - 1][n - 1];
+        return static_cast<int>(dp[m - 1][n - 1]);
     }
 };
